refactor: extract shared icons and deduplicate CSS styles
- Extract CheckIcon, ArrowIcon, NoiseOverlay into SharedIcons.tsx
- Remove duplicate icon definitions from 5 component files
- Consolidate repeated h3/p typography styles in homepage.css
