+++ src/components/Homepage/CloudSection.tsx
@@ -1,20 +1,5 @@
 import React, {useState, useEffect, type ReactNode} from 'react';
-
-function CheckIcon(): ReactNode {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" aria-hidden="true" className="cloud-check-icon">
-      <path fillRule="evenodd" d="M16.704 4.153a.75.75 0 0 1 .143 1.052l-8 10.5a.75.75 0 0 1-1.127.075l-4.5-4.5a.75.75 0 0 1 1.06-1.06l3.894 3.893 7.48-9.817a.75.75 0 0 1 1.05-.143Z" clipRule="evenodd" />
-    </svg>
-  );
-}
-
-function ArrowIcon(): ReactNode {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" className="cloud-arrow-icon">
-      <path fillRule="evenodd" d="M5.22 14.78a.75.75 0 0 0 1.06 0l7.22-7.22v5.69a.75.75 0 0 0 1.5 0v-7.5a.75.75 0 0 0-.75-.75h-7.5a.75.75 0 0 0 0 1.5h5.69l-7.22 7.22a.75.75 0 0 0 0 1.06Z" clipRule="evenodd" />
-    </svg>
-  );
-}
+import {CheckIcon, ArrowIcon} from './SharedIcons';
 
 function IdlingIcon(): ReactNode {
   return (
@@ -257,11 +242,11 @@ export default function CloudSection(): ReactNode {
               </p>
               <ul className="cloud-feature-list">
                 <li>
-                  <CheckIcon />
+                  <CheckIcon className="cloud-check-icon" />
                   <span>Full control via dashboard or CLI</span>
                 </li>
                 <li>
-                  <CheckIcon />
+                  <CheckIcon className="cloud-check-icon" />
                   <span>Instantly add databases, workers,{' '}<br className="tablet-br" />cache, and storage</span>
                 </li>
               </ul>
@@ -287,11 +272,11 @@ export default function CloudSection(): ReactNode {
               </p>
               <ul className="cloud-feature-list">
                 <li>
-                  <CheckIcon />
+                  <CheckIcon className="cloud-check-icon" />
                   <span>Integrates seamlessly with GitHub{' '}<br className="tablet-br" />Actions</span>
                 </li>
                 <li>
-                  <CheckIcon />
+                  <CheckIcon className="cloud-check-icon" />
                   <span>Test migrations and heavy changes{' '}<br className="tablet-br" />safely</span>
                 </li>
               </ul>
