+++ src/components/Homepage/EventsSection.tsx
@@ -1,12 +1,5 @@
 import React, {useState, useRef, useCallback, useEffect, type ReactNode} from 'react';
-
-function ArrowIcon(): ReactNode {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" className="cloud-arrow-icon">
-      <path fillRule="evenodd" d="M5.22 14.78a.75.75 0 0 0 1.06 0l7.22-7.22v5.69a.75.75 0 0 0 1.5 0v-7.5a.75.75 0 0 0-.75-.75h-7.5a.75.75 0 0 0 0 1.5h5.69l-7.22 7.22a.75.75 0 0 0 0 1.06Z" clipRule="evenodd" />
-    </svg>
-  );
-}
+import {ArrowIcon} from './SharedIcons';
 
 const events = [
   {
