+++ src/components/Homepage/FrameworkSection.tsx
@@ -1,13 +1,6 @@
 import React, {useState, useCallback, type ReactNode} from 'react';
+import {CheckIcon} from './SharedIcons';
 import './homepage.css';
-
-function CheckIcon(): ReactNode {
-  return (
-    <svg viewBox="0 0 20 20" fill="currentColor" width="18" height="18" className="check-icon" aria-hidden="true">
-      <path fill="currentColor" fillRule="evenodd" d="M16.704 4.153a.75.75 0 0 1 .143 1.052l-8 10.5a.75.75 0 0 1-1.127.075l-4.5-4.5a.75.75 0 0 1 1.06-1.06l3.894 3.893 7.48-9.817a.75.75 0 0 1 1.05-.143Z" clipRule="evenodd" />
-    </svg>
-  );
-}
 
 function LaravelFileIcon(): ReactNode {
   return (
